feat(studio): fold the chat playground into Studio as a fourth modality
One "try your key" surface instead of two. The standalone /playground chat
debugger is merged into the Studio shell as a first-class peer tab
(Chat · Image · Video · Compare), landing on Chat by default (zero-cost +
near-universal key support → friendliest first touch). The duplicated key
selector / bootstrap / gateway-base resolution collapse into the existing
Studio shell. Net -113 lines while adding a peer modality.

- ChatStudio.vue (new): chat logic migrated from PlaygroundView into a child
  matching ImageStudio's shape; shell passes apiKey/gatewayBase/availableIds,
  the chat model list is derived from the already-probed pool (no extra
  /v1/models fetch). Keeps the external-client integrations + system prompt +
  token usage. No cost-gate (chat is metered after the fact, as before).
- MediaStudioView: leftmost Chat tab + default view, ?mode=chat|image|video|
  bakeoff deep-link, key picker generalized to chat via groupServes /
  PickerModality (modalityForModel) — image/video keep modalityHasTiers.
- /playground now 301-redirects to /studio?mode=chat; sidebar drops the
  Playground entry; i18n studio.chat.* added + cross-modal hero, dead
  playground.* block removed.
- Delete PlaygroundView.vue (TK-only; absent from upstream/main, so no §5.x
  upstream-deletion justification needed).
- frontend-tk sentinel registry updated in lockstep: playground anchors
  migrated to ChatStudio + studio.chat, new studio-mode-chat tab anchored,
  dead playground media-string anchors removed.

Verification: pnpm typecheck + lint:check clean; make test-frontend-critical
green (7 files / 90 tests); mediaTiers/playground/clientIntegrations/
studioDownload specs green (55 tests, incl. new chat/groupServes); production
build clean; frontend-tk + registry-update + upstream-override-marker gates
all pass.

upstream-touch-guarded
sentinel-registry-reviewed
