+++ frontend/src/i18n/locales/en.ts
@@ -98,75 +98,10 @@ export default {
     }
   },
 
-  playground: {
-    title: 'Playground',
-    subtitle: 'Pick an API key first — its group decides which models you can call. Requests go straight to /v1/chat/completions.',
-    apiKey: 'API Key',
-    pickKeyPlaceholder: 'Select an API key…',
-    loadingModels: 'Loading models…',
-    defaultGroup: 'Default group',
-    integrationsTitle: 'Connect external apps',
-    integrationsHint: 'Import the gateway address and the selected API key into a client with one click.',
-    integrationsAppHint: 'Requires the client app installed (opens a custom URL scheme).',
-    copyBaseUrl: 'Copy base URL',
-    copyKey: 'Copy API key',
-    baseUrlCopied: 'Base URL copied',
-    keyCopied: 'API key copied',
-    model: 'Model',
-    temperature: 'Temperature',
-    maxTokens: 'Max tokens',
-    systemPrompt: 'System prompt (optional)',
-    inputPlaceholder: 'Type a message… (Enter to send)',
-    send: 'Send',
-    sending: 'Sending…',
-    cancel: 'Cancel',
-    clear: 'Clear chat',
-    emptyHint: 'Send a message to see the assistant reply here.',
-    manageKeys: 'Manage API keys',
-    mediaInStudio: 'Image / video → Studio',
-    noApiKey: 'No active API key found. Create one under API Keys.',
-    noModels: 'No models returned for your key. Check group routing or try another key.',
-    loadFailed: 'Could not load playground.',
-    pickModelPlaceholder: 'Select a model…',
-    roleUser: 'You',
-    roleAssistant: 'Assistant',
-    avatarUser: 'Me',
-    avatarAssistant: 'AI',
-    limitsHint: 'Up to {turns} turns in memory; max output tokens capped at {maxTok}; 60s timeout per request.',
-    lastUsage: 'Last response usage',
-    promptTokens: 'Prompt tokens',
-    completionTokens: 'Completion tokens',
-    totalTokens: 'Total tokens',
-    cancelled: 'Request cancelled.',
-    requestFailed: 'Request failed.',
-    modalityImage: 'Image generation',
-    modalityVideo: 'Video generation',
-    imageSize: 'Size',
-    imageSizeAuto: 'Auto',
-    imageEmptyHint: 'Describe an image and generate; results show here.',
-    imagePromptPlaceholder: 'Describe the image… (Enter to generate)',
-    imageGenerate: 'Generate image',
-    imageGenerating: 'Generating…',
-    imageNoResult: 'The gateway returned no image data.',
-    videoDuration: 'Duration (s)',
-    videoEmptyHint: 'Describe a video and submit; the task is polled here until it finishes.',
-    videoPromptPlaceholder: 'Describe the video… (Enter to submit)',
-    videoGenerate: 'Generate video',
-    videoSubmitting: 'Submitting…',
-    videoElapsed: 'waited {s}s',
-    videoStopPolling: 'Stop polling',
-    videoRaw: 'Raw response',
-    videoNoUrlHint: 'Task succeeded but no video URL could be parsed — check the raw response.',
-    videoNoTaskId: 'Submit returned no task id.',
-    videoStatusProcessing: 'Processing',
-    videoStatusSucceeded: 'Succeeded',
-    videoStatusFailed: 'Failed'
-  },
-
-  // Media Studio (image / video generation) — TK
+  // Studio (chat / image / video) — TK
   studio: {
     title: 'Studio',
-    subtitle: "Turn what's in your head into images and video you can watch.",
+    subtitle: 'Try everything your key can do — chat, images, and video, all in one place.',
     balance: 'Balance',
     apiKey: 'API Key',
     pickKeyPlaceholder: 'Select an API key…',
@@ -175,6 +110,7 @@ export default {
     noApiKey: 'No active API key found. Create one under API Keys.',
     manageKeys: 'Manage API keys',
     defaultGroup: 'Default group',
+    modeChat: 'Chat',
     modeImage: 'Image',
     modeVideo: 'Video',
     modeBakeoff: 'Bake-Off',
@@ -282,6 +218,38 @@ export default {
       notifyTitle: 'TokenKey Studio',
       notifyEnabled: 'We will notify you when videos finish.',
       notifyDenied: 'Notifications are blocked in your browser.'
+    },
+    chat: {
+      model: 'Model',
+      temperature: 'Temperature',
+      maxTokens: 'Max tokens',
+      pickModelPlaceholder: 'Select a model…',
+      noModels: 'No chat models for this key. Check group routing or pick another key.',
+      systemPrompt: 'System prompt (optional)',
+      inputPlaceholder: 'Type a message… (Enter to send)',
+      send: 'Send',
+      sending: 'Sending…',
+      cancel: 'Cancel',
+      clear: 'Clear chat',
+      emptyHint: 'Send a message to see the assistant reply here.',
+      roleUser: 'You',
+      roleAssistant: 'Assistant',
+      avatarUser: 'Me',
+      avatarAssistant: 'AI',
+      limitsHint: 'Up to {turns} turns in memory; max output tokens capped at {maxTok}; 60s timeout per request.',
+      lastUsage: 'Last response usage',
+      promptTokens: 'Prompt tokens',
+      completionTokens: 'Completion tokens',
+      totalTokens: 'Total tokens',
+      cancelled: 'Request cancelled.',
+      requestFailed: 'Request failed.',
+      integrationsTitle: 'Connect external apps',
+      integrationsHint: 'Import the gateway address and the selected API key into a client with one click.',
+      integrationsAppHint: 'Requires the client app installed (opens a custom URL scheme).',
+      copyBaseUrl: 'Copy base URL',
+      copyKey: 'Copy API key',
+      baseUrlCopied: 'Base URL copied',
+      keyCopied: 'API key copied'
     }
   },
 
@@ -696,7 +664,6 @@ export default {
     paymentDashboard: 'Payment Dashboard',
     paymentConfig: 'Payment Config',
     paymentPlans: 'Plans',
-    playground: 'Playground',
     studio: 'Studio',
     channelManagement: 'Channels',
     channelPricing: 'Channel Pricing',
