+++ frontend/src/i18n/locales/zh.ts
@@ -96,75 +96,10 @@ export default {
     }
   },
 
-  playground: {
-    title: 'Playground',
-    subtitle: '先选择 API 密钥——密钥所属分组决定可用模型；请求直达 /v1/chat/completions。',
-    apiKey: 'API 密钥',
-    pickKeyPlaceholder: '请选择 API 密钥…',
-    loadingModels: '加载模型中…',
-    defaultGroup: '默认分组',
-    integrationsTitle: '一键接入外部客户端',
-    integrationsHint: '将网关地址与当前选中的 API 密钥一键导入客户端。',
-    integrationsAppHint: '需已安装对应客户端（通过自定义 URL Scheme 打开）。',
-    copyBaseUrl: '复制接入地址',
-    copyKey: '复制 API 密钥',
-    baseUrlCopied: '接入地址已复制',
-    keyCopied: 'API 密钥已复制',
-    model: '模型',
-    temperature: '温度',
-    maxTokens: 'Max tokens',
-    systemPrompt: '系统提示词（可选）',
-    inputPlaceholder: '输入消息…（Enter 发送）',
-    send: '发送',
-    sending: '发送中…',
-    cancel: '取消',
-    clear: '清空会话',
-    emptyHint: '发送一条消息，助手回复将显示在此。',
-    manageKeys: '管理 API 密钥',
-    mediaInStudio: '图片 / 视频 → 工作室',
-    noApiKey: '未找到可用的 API 密钥，请先在「API 密钥」页面创建。',
-    noModels: '未返回可用模型，请检查分组路由或更换密钥。',
-    loadFailed: '无法加载 Playground。',
-    pickModelPlaceholder: '请选择模型…',
-    roleUser: '你',
-    roleAssistant: '助手',
-    avatarUser: '我',
-    avatarAssistant: 'AI',
-    limitsHint: '浏览器内最多保留 {turns} 轮对话；单次输出上限 {maxTok} tokens；单次请求超时 60 秒。',
-    lastUsage: '上次响应用量',
-    promptTokens: '输入 tokens',
-    completionTokens: '输出 tokens',
-    totalTokens: '总计 tokens',
-    cancelled: '请求已取消。',
-    requestFailed: '请求失败。',
-    modalityImage: '图片生成',
-    modalityVideo: '视频生成',
-    imageSize: '尺寸',
-    imageSizeAuto: '自动',
-    imageEmptyHint: '输入图片描述并生成，结果将显示在此。',
-    imagePromptPlaceholder: '描述要生成的图片…（Enter 生成）',
-    imageGenerate: '生成图片',
-    imageGenerating: '生成中…',
-    imageNoResult: '网关未返回图片数据。',
-    videoDuration: '时长（秒）',
-    videoEmptyHint: '输入视频描述并提交，任务会在此轮询直到完成。',
-    videoPromptPlaceholder: '描述要生成的视频…（Enter 提交）',
-    videoGenerate: '生成视频',
-    videoSubmitting: '提交中…',
-    videoElapsed: '已等待 {s} 秒',
-    videoStopPolling: '停止轮询',
-    videoRaw: '原始响应',
-    videoNoUrlHint: '任务成功但未解析出视频链接，请查看原始响应。',
-    videoNoTaskId: '提交未返回任务 ID。',
-    videoStatusProcessing: '处理中',
-    videoStatusSucceeded: '已完成',
-    videoStatusFailed: '失败'
-  },
-
-  // 媒体工作室（图片 / 视频生成）— TK
+  // 工作室（对话 / 图片 / 视频）— TK
   studio: {
     title: '工作室',
-    subtitle: '把脑子里的画面，变成能播放的图片和视频。',
+    subtitle: '你的密钥能做的，都在这里试——对话、图片、视频，一处搞定。',
     balance: '余额',
     apiKey: 'API 密钥',
     pickKeyPlaceholder: '选择一个 API 密钥…',
@@ -173,6 +108,7 @@ export default {
     noApiKey: '没有可用的 API 密钥，请先在「API 密钥」里创建一个。',
     manageKeys: '管理 API 密钥',
     defaultGroup: '默认分组',
+    modeChat: '对话',
     modeImage: '图片',
     modeVideo: '视频',
     modeBakeoff: '同台对比',
@@ -280,6 +216,38 @@ export default {
       notifyTitle: 'TokenKey 工作室',
       notifyEnabled: '视频完成时会通知你。',
       notifyDenied: '浏览器已屏蔽通知。'
+    },
+    chat: {
+      model: '模型',
+      temperature: '温度',
+      maxTokens: 'Max tokens',
+      pickModelPlaceholder: '请选择模型…',
+      noModels: '该密钥无可用对话模型，请检查分组路由或更换密钥。',
+      systemPrompt: '系统提示词（可选）',
+      inputPlaceholder: '输入消息…（Enter 发送）',
+      send: '发送',
+      sending: '发送中…',
+      cancel: '取消',
+      clear: '清空会话',
+      emptyHint: '发送一条消息，助手回复将显示在此。',
+      roleUser: '你',
+      roleAssistant: '助手',
+      avatarUser: '我',
+      avatarAssistant: 'AI',
+      limitsHint: '浏览器内最多保留 {turns} 轮对话；单次输出上限 {maxTok} tokens；单次请求超时 60 秒。',
+      lastUsage: '上次响应用量',
+      promptTokens: '输入 tokens',
+      completionTokens: '输出 tokens',
+      totalTokens: '总计 tokens',
+      cancelled: '请求已取消。',
+      requestFailed: '请求失败。',
+      integrationsTitle: '一键接入外部客户端',
+      integrationsHint: '将网关地址与当前选中的 API 密钥一键导入客户端。',
+      integrationsAppHint: '需已安装对应客户端（通过自定义 URL Scheme 打开）。',
+      copyBaseUrl: '复制接入地址',
+      copyKey: '复制 API 密钥',
+      baseUrlCopied: '接入地址已复制',
+      keyCopied: 'API 密钥已复制'
     }
   },
 
@@ -694,7 +662,6 @@ export default {
     paymentDashboard: '支付概览',
     paymentConfig: '支付配置',
     paymentPlans: '订阅套餐',
-    playground: 'Playground',
     studio: '工作室',
     channelManagement: '渠道管理',
     channelPricing: '渠道定价',
